Steam - List AppManifest Items
For identifying unneeded .acf items in the SteamLibrary\steamapps directory. Removing them prevents steam trying to update a game that has been moved or deleted through FileExplorer instead of the Steam Application itself.

The script's state is Very Basic. Run the script, select your steamapps folder and the items will be listed in the powershell window. You have to manually go into steamapps and delete the entries that are currently not installed or in directory.